Allowed bootstrap and (especially) animate to track NG version
Bootstrap and bootstrap-css generally have no problems tracking the latest version of one another, however we need them to track the version of angular that is selected.  Similarly, ng-animate needs to track the version of angular exactly since ng-animate tends to break with the most minor version differences between it and angular.