Updated gulpfile for more keywords, added repository key
Installing angular-redhawk was producing a warning we forgot to
put the 'repository' key in place, so it has been added.  The
gulpfile was also updated to migrate repository and bugs to the
distributed package.json as well.

The remaining warning about ts-simple-ast (dependency of compodoc,
only used in development) is because they have their TypeScript
dependency set to "2.6.1" rather than something more relaxed.  It
still works fine on 2.5.3, so it can be ignored.