Bug fix for when default values are not provided in the property
Some properties are not defined (in the PRF) with values, so the
parsing to JSON results in `undefined` or `null`.  If you then
configure/allocate/deallocate with that structure, the REST
endpoint callback into the REDHAWK python interface will emit CORBA
warnings related to it having been passed TC_null members when the
property is not defined that way.